UI/mainbar.js - split and roll up
This splits the mainbar.js into various files and bundles them for distribution via rollup according to the new approach for js-code introduced by @alex40724 (thx!!). This also introduces some tests on the mainbar and runs them. This make the mainbar.js look a lot tidier then before.
