+++ src/UI/templates/js/MainControls/rollup.config.js
@@ -1,0 +1,7 @@
+export default {
+  input: './src/mainbar.js',
+  output: {
+    file: './dist/mainbar.js',
+    format: 'es'
+  }
+};
