+++ test/mainbar.test.js
@@ -1,0 +1,1 @@
+../src/UI/templates/js/MainControls/test/mainbar.test.js
